Previous fixes for #92 and #123 (#157)
* Fix bot disabling issue #92 again

Disable the acceleration when turning off the bot.
Improved formatting (tabs and spaces).

* Fix mobile render issue #123 again

This fixes the mobile rendering bug, and simplifies the function to be
more in line with the others and to save the mobileRender variable.

* Remove space before function parentheses.

As requested by Hound. Since this just serves to clear the onmousemove()-function
(at least I think it does), what should the function be named since Hound asks for this as well.

* Spacing error

Removes some excessive spaces.